LayoutTests/storage/indexeddb/modern/opendatabase-request-event-expected.txt

ALERT: [object IDBOpenDBRequest]
ALERT: error [object Event]
ALERT: Done
This test calls open on window.indexedDB in various ways, verifying that IDBOpenDBRequest objects are returned when expected and exceptions are thrown when expected.
